Fix wrong variable type in method declaration

+++ app/src/main/java/com/jekz/stepitup/model/step/IntervalStepCounter.java
@@ -5,7 +5,7 @@ package com.jekz.stepitup.model.step;
  */
 
 public interface IntervalStepCounter extends StepCounter {
-    void setInterval(int numOfSeconds);
+    void setInterval(long numOfSeconds);
 
     void notifySessionEnded(Session session);
